requirements.txt
playwright>=1.48.0
python-docx>=1.1.0
PyPDF2>=3.0.1
pandas>=2.2.0
httpx>=0.27.0
anthropic>=0.39.0
python-dotenv>=1.0.1
